feat(sdk): split node-only exports into sdk/node
