+++ packages/sdk/src/node/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @fiber-pay/sdk/node
+ * Node-focused SDK entry point.
+ *
+ * Includes all universal APIs from the root entry plus Node-only
+ * server integrations such as the L402 middleware module.
+ *
+ * @packageDocumentation
+ */
+
+export * from '../index.js';
+export * from '../l402/index.js';
