+++ packages/sdk/tsup.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: {
     index: 'src/index.ts',
+    'node/index': 'src/node/index.ts',
     'browser/index': 'src/browser/index.ts',
   },
   format: ['esm'],
